commands: move show code to show.rs
